Fixed a possible PropertyOrFieldNotInitializedException
In some cases the "HeaderLayout" property will not be set and cause a
"PropertyOrFieldNotInitializedException" when accessing it later on.
Loading the existing value fixes this and will (in those cases) apply
the already set value again.

Loading the web URL is also not required, because "ExecutePostAsync"
will ensure it again.