Fix "The request was already sent" error when getting app metadata
When getting app metadata after an app has been added, if the metadataResponse.StatusCode == HttpStatusCode.NotFound then it will retry the request but without changing the HttpRequestMessage. This causes the following error "System.InvalidOperation: The request was already sent. Cannot send the same request multiple times".  
In this fix I included the creation of the HttpRequestMessage inside the while loop so that each time a retry is made, it creates a new HttpRequestMessage.